-I Just finished the Template For MAge. Mage took much longer cause There appears to be two types of mages. The SupportMAge and the one thats not. Both have diferent values for the variables and even the way they level up. So I needed to write the logic for what happens in either case over what I did previously
-I of course previously did the template for RangeCharacters but that was actually the specs for Archer.
There was no provided specs for Range.
This means the Range has different vaules for the variables AND DIFFERENT METHODS because the archer and Mage have unique ways of fightin and leveling up.
-But sinec they didn't give the actually Range values ill leave it as is.

I also want to take the time to write the questions we have to ask Lizza and Alex here.
||What does the paw mean? Is it VC?||

||There are nine Cards, could this be the use of Mana?||

||Whats the Orange Bar over the Lion?||

||What happens between Level 1-3 for Mage and Archers?||