-I tried to make activationrequirement( object classInstance) work but since the object instance doesnt exist yet it seems it cant access it properties.
Thats a problem cause alot of how i imagined items will work depended on that.

An alternative was thought up though. What will happen is that instances of an item (relic) will be instanciated within CharacterPersona. Then that will be accessed by the templates. How that items method will work on a given template is a bit foggy in my mind.

- I introduced a bool called owner. It was supposed to be the object to equate the argument 'object CharacterInstance' that would be used to acesss an instance properties. But that didn't pan out well.

-So because the owner object wasn't working as planned i just had the effect of the relic happen on every possible templatetype instance so that if called in a template it will respond accordingly. I just hope it doesn't  work on every made instance. But i doubt it cause of scope. ( but this is NOT guarranteed)