-It was really difficult but i managed to get the position of the target behind the target. All the code is in the first card switch section. I did it! -I essentially created placeholders of the posistions, gameobjects and the Persona component of the targetbehind.
Then i sift throught the postions of playing characters and check to see if it is the same as the target parameters position.
Then i take the posistion of the confirmed value and subtract the vector to get to the next position over.
After that, i made a second logic to get the character in that position.
Then i GetComponent<Persona> of that GameObject. Then use that in the Physical damage()